Removed reduandant execution method and moved server_mlmd_pull inside get_data.py file
