+++ server/app/main.py
@@ -21,8 +21,8 @@ from server.app.get_data import (
     get_all_exe_ids,
     async_api,
     get_model_data,
-    executions_list
-
+    executions_list,
+    server_mlmd_pull
 )
 from server.app.query_execution_lineage_d3tree import query_execution_lineage_d3tree
 from server.app.query_artifact_lineage_d3tree import query_artifact_lineage_d3tree
@@ -202,19 +202,6 @@ async def execution(request: Request,
 
     """Retrieve paginated executions with filtering, sorting, and full-text search."""
     return await fetch_executions(db, pipeline_name, filter_value, active_page, record_per_page, sort_field, sort_order)
-    
-
-# api to display executions available in mlmd file[from postgres]
-@app.get("/executions/{pipeline_name}")
-async def execution(request: Request,
-                   pipeline_name: str,
-                   active_page: int = Query(1, description="Page number", gt=0),
-                   filter_value: str = Query("", description="Search based on value"),  
-                   sort_order: str = Query("asc", description="Sort by context_type(asc or desc)"),
-                   db: AsyncSession = Depends(get_db)
-                   ):
-    """Retrieve paginated executions with filtering, sorting, and full-text search."""
-    return await fetch_executions(db, pipeline_name, filter_value, active_page, 5, sort_order)
     
 
 @app.get("/execution-lineage/tangled-tree/{uuid}/{pipeline_name}")
@@ -409,7 +396,6 @@ async def upload_label(request:Request, file: UploadFile = File(..., description
         if os.path.exists(file_path):
             return {"message": f"File '{file.filename}' already exists at {labels_dir}. Skipping upload."}
 
-
         # Save the uploaded file
         with open(file_path, "wb") as buffer:
             buffer.write(await file.read())
@@ -448,7 +434,6 @@ async def get_label_data(file_name: str) -> str:
         return content
     except Exception as e:
         raise HTTPException(status_code=500, detail=f"Error reading file: {str(e)}")
-
 
 
 @app.post("/register-server")
@@ -492,88 +477,6 @@ async def acknowledge(request: AcknowledgeRequest):
     return {
         "message": f"Hi, I acknowledge your request.",
     }
-
-
-async def server_mlmd_pull(server_url, last_sync_time):
-    """
-    Fetch mlmd data from a specified server.
-
-    Args:
-        server_url (str): The full URL of the target server.
-        last_sync_time (int): The last sync time in milliseconds since epoch.
-
-    Returns:
-        dict: The mlmd data fetched from the specified server.
-
-    Raises:
-        HTTPException: If the server is not reachable or an error occurs during the request.
-    """
-    try:
-        # Step 1: Send a request to the target server to fetch mlmd data
-        async with httpx.AsyncClient(timeout=300.0) as client:
-            try:
-                response = await client.post(f"{server_url}/api/mlmd_pull", json={'last_sync_time': last_sync_time})
-
-                if response.status_code != 200:
-                    raise HTTPException(status_code=500, detail="Target server did not respond successfully")
-
-                json_payload = response.json()
-                python_env_store_path = "/cmf-server/data/env"
-
-                if last_sync_time:
-                    # Extract the Environment file names from the JSON payload
-                    data = json_payload
-                    jsonpath_expr = parse('$..events[?(@.artifact.type == "Environment")].artifact.name')
-                    environment_names = {match.value.split(":")[0].split("/")[-1] for match in jsonpath_expr.find(data)}
-
-                    # Check if the list is empty or not
-                    if len(environment_names) == 0: 
-                        print("No Environment files are found inside json payload.")
-                        return json_payload
-
-                    list_of_files = list(environment_names)
-                    python_env_zip = await client.get(f"{server_url}/api/download-python-env", params=list_of_files)
-                else:
-                    python_env_zip = await client.get(f"{server_url}/api/download-python-env", params=None)
-
-                if python_env_zip.status_code == 200:
-                    try:
-                        # Create the directory if it doesn't exist
-                        os.makedirs(python_env_store_path, exist_ok=True)
-
-                        # Unzip the zip file content
-                        with zipfile.ZipFile(io.BytesIO(python_env_zip.content)) as zf:
-                            # Extract all files to a temporary directory
-                            temp_dir = os.path.join(python_env_store_path, "temp_extracted")
-                            os.makedirs(temp_dir, exist_ok=True)
-                            zf.extractall(temp_dir)
-
-                            # Move all extracted files to the target directory
-                            for root, dirs, files in os.walk(temp_dir):
-                                for file in files:
-                                    src_file = os.path.join(root, file)
-                                    dest_file = os.path.join(python_env_store_path, file)
-                                    try:
-                                        os.rename(src_file, dest_file)
-                                        print(f"Moved {src_file} to {dest_file}")
-                                    except Exception as e:
-                                        print(f"Failed to move {src_file} to {dest_file}: {e}")
-
-                            # Clean up the temporary directory
-                            os.rmdir(temp_dir)
-                        # print("Storing at:", os.path.abspath(python_env_store_path))
-                        # print("Files in target directory:", os.listdir(python_env_store_path))
-                        print("All files stored successfully.")
-                    except Exception as e:
-                        print(f"Error during file extraction or storage: {e}")
-                else:
-                    print(f"Failed to download ZIP file. Status code: {python_env_zip.status_code}")
-            except httpx.RequestError:
-                raise HTTPException(status_code=500, detail="Target server is not reachable")
-        return json_payload
-
-    except Exception as e:
-        raise HTTPException(status_code=500, detail=f"Failed to fetch mlmd data: {e}")
 
 
 @app.post("/sync")
